# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/roles": {
      "post": {
        "operationId": "RolesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Role created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        },
        "summary": "Create a new role",
        "tags": ["roles"]
      },
      "get": {
        "operationId": "RolesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all roles",
        "tags": ["roles"]
      }
    },
    "/api/roles/{id}": {
      "get": {
        "operationId": "RolesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        },
        "summary": "Get a role by ID",
        "tags": ["roles"]
      },
      "put": {
        "operationId": "RolesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Partial role data to update",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        },
        "summary": "Update a role by ID",
        "tags": ["roles"]
      },
      "delete": {
        "operationId": "RolesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Role ID",
            "schema": {
              "example": "507f1f77bcf86cd799439011",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Role deleted successfully"
          },
          "404": {
            "description": "Role not found"
          }
        },
        "summary": "Delete a role by ID",
        "tags": ["roles"]
      }
    },
    "/api/roles/name/{name}": {
      "get": {
        "operationId": "RolesController_findByName",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Role name",
            "schema": {
              "example": "admin",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Role found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          }
        },
        "summary": "Get a role by name",
        "tags": ["roles"]
      }
    },
...
    "/api/tickets": {
      "post": {
        "operationId": "TicketingController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      },
      "get": {
        "operationId": "TicketingController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      }
    },
    "/api/tickets/{id}": {
      "get": {
        "operationId": "TicketingController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      },
      "put": {
        "operationId": "TicketingController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTicketDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      },
      "delete": {
        "operationId": "TicketingController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      }
    },
    "/api/tickets/{id}/status": {
      "get": {
        "operationId": "TicketingController_findStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      },
      "put": {
        "operationId": "TicketingController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      }
    },
    "/api/tickets/{id}/escalate": {
      "post": {
        "operationId": "TicketingController_escalateTicket",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Ticketing"]
      }
    },
...
    "/api/transaction": {
      "post": {
        "operationId": "TransactionController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Transaction"]
      },
      "get": {
        "operationId": "TransactionController_findAll",
        "parameters": [
          {
            "name": "terminalId",
            "required": true,
            "in": "query",
            "description": "Zarinpal terminal ID",
            "schema": {
              "example": "a3c16110-f184-44e2-ad26-649387845a94",
              "type": "string"
            }
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filter for transactions",
            "schema": {
              "example": "all",
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": true,
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "example": 0,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "description": "Limit for pagination",
            "schema": {
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Transaction"]
      }
    },
    "/api/transaction/inquire": {
      "post": {
        "operationId": "TransactionController_inquire",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InquireZarinpalTransactionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": ["Transaction"]
      }
    },
...
  },
  "info": {
    "title": "practice",
    "description": "API documentation for my practice app",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Permission": {
        "type": "object",
        "properties": {
          "resource": {
            "type": "string",
            "enum": ["carts", "", "", "", "", "", "", "", "", "", "", "", "all"],
            "example": ""
          },
          "actions": {
            "type": "array",
            "example": ["read", "update"],
            "items": {
              "type": "string",
              "enum": ["read", "update", "create", "delete", "manage"]
            }
          }
        },
        "required": ["resource", "actions"]
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "admin"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        },
        "required": ["name", "permissions"]
      },
      "Role": {
        "type": "object",
        "properties": {}
      },
...
      "CreateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the ticket",
            "example": "Issue with payment"
          },
          "description": {
            "type": "string",
            "description": "Description of the ticket",
            "example": "Payment failed during checkout"
          },
          "status": {
            "type": "string",
            "description": "Status of the ticket",
            "enum": ["open", "in_progress", "resolved", "closed", "reopened", "escalated"],
            "default": "open"
          },
          "priority": {
            "type": "string",
            "description": "Priority of the ticket",
            "enum": ["low", "medium", "high", "urgent"],
            "default": "low"
          },
          "createdBy": {
            "type": "string",
            "description": "ID of the user who created the ticket",
            "example": "507f1f77bcf86cd799439011"
          },
          "assignedTo": {
            "type": "string",
            "description": "ID of the user assigned to the ticket",
            "example": "507f1f77bcf86cd799439012"
          }
        },
        "required": ["title", "description"]
      },
      "UpdateTicketDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title of the ticket",
            "example": "Updated issue with payment"
          },
          "description": {
            "type": "string",
            "description": "Description of the ticket",
            "example": "Updated description for payment issue"
          },
          "status": {
            "type": "string",
            "description": "Status of the ticket",
            "enum": ["open", "in_progress", "resolved", "closed", "reopened", "escalated"],
            "example": "in_progress"
          },
          "priority": {
            "type": "string",
            "description": "Priority of the ticket",
            "enum": ["low", "medium", "high", "urgent"],
            "example": "high"
          },
          "assignedTo": {
            "type": "string",
            "description": "ID of the user assigned to the ticket",
            "example": "507f1f77bcf86cd799439012"
          }
        }
      },
...
      "CreateTransactionDto": {
        "type": "object",
        "properties": {}
      },
      "InquireZarinpalTransactionDto": {
        "type": "object",
        "properties": {
          "authority": {
            "type": "string",
            "description": "Transaction authority",
            "example": "A00000000000000000000000000000000001"
          }
        },
        "required": ["authority"]
      },
...
    }
  }
}